+++ database/migrations/2023_07_04_095858_create_referrals_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('referral_id');
             $table->foreignId('referee_id');
-            $table->integer('amount')->default(0);
+            $table->decimal('amount')->default(0);
             $table->boolean('status')->default(true);
             $table->timestamps();
         });
